requirements.txt
python-dotenv
requests
langchain
langchain-pinecone
langchain-voyageai